fix mn-number webcomponent and files
